final validation

+++ server.js
@@ -255,7 +255,7 @@ app.get("/home/search",async function(req,res){
 //Profile
 app.get("/profile/profile",async function(req,res){
     //Get reviews, threads, comments
-    if (typeof(req.session.user.email) == 'undefined'){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'||typeof(req.session.user.email) == 'undefined'){res.sendStatus(401);return;};
     let profile = {email:req.session.user.email,reviews:[],threads:[],comments:[]};
     profile.reviews = await db.collection("reviews").find({authorid:req.session.user._id}).toArray();
     profile.threads = await db.collection("threads").find({authorid:req.session.user._id}).toArray();
@@ -263,41 +263,41 @@ app.get("/profile/profile",async function(req,res){
     res.send(profile);
 });
 app.delete("/profile/delete/review",async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let result = await db.collection("reviews").deleteOne({_id:new Mongo.ObjectID(req.query.id)});
     res.send(200);
 });
 app.delete("/profile/delete/thread", async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let result = await db.collection("threads").deleteOne({_id:new Mongo.ObjectID(req.query.id)});
     res.send(200);
 });
 app.delete("/profile/delete/comment", async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let result = await db.collection("comments").deleteOne({_id:new Mongo.ObjectID(req.query.id)});
     res.send(200);
 });
 //Profile Edit
 app.get("/profileedit/profile",function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let profileEdit = {email:req.session.user.email,password1:req.session.user.password,password2:""}
     res.send(JSON.stringify(profileEdit));
 });
 app.post("/profileedit/profile/edit",async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     await db.collection("profiles").updateOne({_id:new Mongo.ObjectID(req.session.user._id)},{email:req.body.params.profile.email,password:req.body.params.profile.password1},{upsert:true})
     req.session.user = await db.collection("profiles").findOne({_id:new Mongo.ObjectID(req.session.user._id)});
     res.send(201);
 });
 //Thread Edit
 app.get("/threadedit/anime",async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let result = await animeNewsNetworkApi.getById([req.session.threadEdit.animeid]);
     res.send(JSON.stringify(result[0]));
 });
 app.get("/threadedit/get", function(req,res){
     //gets thread by id
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     if(req.session.threadEdit.id != null){
         db.collection('threads').findOne({_id:new Mongo.ObjectID(req.session.threadEdit.id)}, function(err, result){
             if (err) throw error
@@ -308,7 +308,7 @@ app.get("/threadedit/get", function(req,res){
     }
 });
 app.post("/threadedit/save", function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     //saves thread
     if (req.session.threadEdit.id){
         req.body.params.thread._id = req.session.threadEdit.id;
@@ -325,12 +325,12 @@ app.post("/threadedit/save", function(req,res){
 });
 //Review Edit
 app.get("/reviewedit/anime",async function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     let result = await animeNewsNetworkApi.getById([req.session.reviewEdit.animeid])
     res.send(JSON.stringify(result[0]));
 });
 app.get("/reviewedit/get", function(req,res){
-    if (!req.session.user){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     //gets review by id
     if(req.session.reviewEdit.id != null){
         db.collection('reviews').findOne({_id:new Mongo.ObjectID(req.session.reviewEdit.id)}, function(err, result){
@@ -342,7 +342,7 @@ app.get("/reviewedit/get", function(req,res){
     }
 });
 app.post("/reviewedit/save",function(req,res){
-    if (!req.session.user){res.sendStatus(401)};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401)};
     //saves review
     if (req.session.reviewEdit.id){
         req.body.params.review.id = req.session.reviewEdit.id;
@@ -370,10 +370,11 @@ app.post('/signup',async function(req,res){
     let exists = await db.collection("profiles").findOne({email:req.body.params.email});
     if (exists){
         res.sendStatus(401);
-    } 
-    let result = await db.collection("profiles").insert({email:req.body.params.email,password:req.body.params.password});
-    updateAdmin({accountsCreated:1});
-    res.send({email:req.body.params.email});
+    } else {
+        let result = await db.collection("profiles").insert({email:req.body.params.email,password:req.body.params.password});
+        updateAdmin({accountsCreated:1});
+        res.send({email:req.body.params.email});
+    }
 });
 app.post("/login", async function(req,res){
     //Login goes here
@@ -436,17 +437,17 @@ app.get("/popup/anime", async function(req,res){
     res.send(JSON.stringify(await anime));
 });
 app.post("/popup/anime/addReview", function(req,res){
-    if (req.session.user==undefined){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     req.session.reviewEdit = {id:null,animeid:req.body.params.id};
     res.send(201);
 });
 app.post("/popup/anime/addThread", function(req,res){
-    if (req.session.user==undefined){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     req.session.threadEdit = {id:null,animeid:req.body.params.id};
     res.send(201);
 })
 app.post("/popup/anime/addComment", function(req,res){
-    if (req.session.user==undefined){res.sendStatus(401);return;};
+    if (typeof(req.session)=='undefined'||typeof(req.session.user)=='undefined'){res.sendStatus(401);return;};
     db.collection("comments").insert({id:req.body.params.id,comment:req.body.params.comment,authorid:req.session.user._id,author:req.session.user.email,date:new Date()});
     updateAdmin({reviewsCreated:1});
     res.send(201);
